+++ packages/library/src/send-and-confirm-durable-nonce-transaction.ts
@@ -1,0 +1,54 @@
+import type { GetAccountInfoApi, GetSignatureStatusesApi, Rpc, SendTransactionApi } from '@solana/rpc';
+import type { AccountNotificationsApi, RpcSubscriptions, SignatureNotificationsApi } from '@solana/rpc-subscriptions';
+import {
+    createNonceInvalidationPromiseFactory,
+    createRecentSignatureConfirmationPromiseFactory,
+    waitForDurableNonceTransactionConfirmation,
+} from '@solana/transaction-confirmation';
+import { FullySignedTransaction, TransactionWithDurableNonceLifetime } from '@solana/transactions';
+
+import { sendAndConfirmDurableNonceTransaction_INTERNAL_ONLY_DO_NOT_EXPORT } from './send-transaction-internal';
+
+type SendAndConfirmDurableNonceTransactionFunction = (
+    transaction: FullySignedTransaction & TransactionWithDurableNonceLifetime,
+    config: Omit<
+        Parameters<typeof sendAndConfirmDurableNonceTransaction_INTERNAL_ONLY_DO_NOT_EXPORT>[0],
+        'confirmDurableNonceTransaction' | 'rpc' | 'transaction'
+    >,
+) => Promise<void>;
+
+interface SendAndConfirmDurableNonceTransactionFactoryConfig {
+    rpc: Rpc<GetAccountInfoApi & GetSignatureStatusesApi & SendTransactionApi>;
+    rpcSubscriptions: RpcSubscriptions<AccountNotificationsApi & SignatureNotificationsApi>;
+}
+
+export function sendAndConfirmDurableNonceTransactionFactory({
+    rpc,
+    rpcSubscriptions,
+}: SendAndConfirmDurableNonceTransactionFactoryConfig): SendAndConfirmDurableNonceTransactionFunction {
+    const getNonceInvalidationPromise = createNonceInvalidationPromiseFactory({ rpc, rpcSubscriptions });
+    const getRecentSignatureConfirmationPromise = createRecentSignatureConfirmationPromiseFactory({
+        rpc,
+        rpcSubscriptions,
+    });
+    async function confirmDurableNonceTransaction(
+        config: Omit<
+            Parameters<typeof waitForDurableNonceTransactionConfirmation>[0],
+            'getNonceInvalidationPromise' | 'getRecentSignatureConfirmationPromise'
+        >,
+    ) {
+        await waitForDurableNonceTransactionConfirmation({
+            ...config,
+            getNonceInvalidationPromise,
+            getRecentSignatureConfirmationPromise,
+        });
+    }
+    return async function sendAndConfirmDurableNonceTransaction(transaction, config) {
+        await sendAndConfirmDurableNonceTransaction_INTERNAL_ONLY_DO_NOT_EXPORT({
+            ...config,
+            confirmDurableNonceTransaction,
+            rpc,
+            transaction,
+        });
+    };
+}
