+++ packages/library/src/send-and-confirm-transaction.ts
@@ -1,0 +1,57 @@
+import type { GetEpochInfoApi, GetSignatureStatusesApi, Rpc, SendTransactionApi } from '@solana/rpc';
+import type { RpcSubscriptions, SignatureNotificationsApi, SlotNotificationsApi } from '@solana/rpc-subscriptions';
+import {
+    createBlockHeightExceedencePromiseFactory,
+    createRecentSignatureConfirmationPromiseFactory,
+    waitForRecentTransactionConfirmation,
+} from '@solana/transaction-confirmation';
+import { FullySignedTransaction, TransactionWithBlockhashLifetime } from '@solana/transactions';
+
+import { sendAndConfirmTransactionWithBlockhashLifetime_INTERNAL_ONLY_DO_NOT_EXPORT } from './send-transaction-internal';
+
+type SendAndConfirmTransactionWithBlockhashLifetimeFunction = (
+    transaction: FullySignedTransaction & TransactionWithBlockhashLifetime,
+    config: Omit<
+        Parameters<typeof sendAndConfirmTransactionWithBlockhashLifetime_INTERNAL_ONLY_DO_NOT_EXPORT>[0],
+        'confirmRecentTransaction' | 'rpc' | 'transaction'
+    >,
+) => Promise<void>;
+
+interface SendAndConfirmTransactionWithBlockhashLifetimeFactoryConfig {
+    rpc: Rpc<GetEpochInfoApi & GetSignatureStatusesApi & SendTransactionApi>;
+    rpcSubscriptions: RpcSubscriptions<SignatureNotificationsApi & SlotNotificationsApi>;
+}
+
+export function sendAndConfirmTransactionFactory({
+    rpc,
+    rpcSubscriptions,
+}: SendAndConfirmTransactionWithBlockhashLifetimeFactoryConfig): SendAndConfirmTransactionWithBlockhashLifetimeFunction {
+    const getBlockHeightExceedencePromise = createBlockHeightExceedencePromiseFactory({
+        rpc,
+        rpcSubscriptions,
+    });
+    const getRecentSignatureConfirmationPromise = createRecentSignatureConfirmationPromiseFactory({
+        rpc,
+        rpcSubscriptions,
+    });
+    async function confirmRecentTransaction(
+        config: Omit<
+            Parameters<typeof waitForRecentTransactionConfirmation>[0],
+            'getBlockHeightExceedencePromise' | 'getRecentSignatureConfirmationPromise'
+        >,
+    ) {
+        await waitForRecentTransactionConfirmation({
+            ...config,
+            getBlockHeightExceedencePromise,
+            getRecentSignatureConfirmationPromise,
+        });
+    }
+    return async function sendAndConfirmTransaction(transaction, config) {
+        await sendAndConfirmTransactionWithBlockhashLifetime_INTERNAL_ONLY_DO_NOT_EXPORT({
+            ...config,
+            confirmRecentTransaction,
+            rpc,
+            transaction,
+        });
+    };
+}
